pyproject.toml
[project]
name = "poetry-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Luciano Lorenti", email = "lucianolorenti@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts>=0.1.3",
    "langdetect>=1.0.9",
    "melotts>=0.1.0",
    "moviepy>=2.2.1",
    "numpy>=1.26.4",
    "onnx==1.18",
    "openpyxl>=3.1.5",
    "pandas>=1.5.3",
    "requests>=2.32.5",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "tqdm>=4.67.1",
    "tts>=0.22.0",
    "typer>=0.21.1",
]

[project.scripts]
poetry-reader = "poetry_reader:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[tool.uv.extra-build-dependencies]
pkuseg = ["numpy"]
